moved encryption
~ moved encryption module to foundation
+ created function to generate shared secrets
